Self Alignment will no longer affect self positioning within a container.
i.e. Node A's alignment will only affect children of Node A, no longer A itself and its children.

i.e. If I make the alignment of a pane CENTER, and the parent container is TOP_LEFT, that centered pane will be in the TOP_LEFT, and its children will be centered.

See PanesInPanes demo.